deps/v8/test/webkit/class-syntax-scoping-expected.txt

Tests for scoping of variables in ES6 class syntax

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test() is "PASS"
PASS successfullyParsed is true

TEST COMPLETE
